+++ packages/mobx-state-tree/test/reference-custom.ts
@@ -16,7 +16,7 @@ test("it should support custom references - basics", () => {
         id: types.identifier,
         name: types.string
     })
-    const UserByNameReference = types.maybe(
+    const UserByNameReference = types.maybeNull(
         types.reference(User, {
             // given an identifier, find the user
             get(identifier /* string */, parent: typeof Store.Type /*Store*/) {